chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` file. It adds the `@liveblocks/client` and `@liveblocks/react` packages, as well as the `fabric` package. It also updates the `next` package to version 14.2.5 and adds the `uuid` package. Additionally, it updates the dev dependencies, including `eslint` and `eslint-config-next`, and updates `typescript` to version 5.