chore: Update Liveblocks configuration and dependencies
This commit updates the Liveblocks configuration in the `liveblocks.config.ts` file. It adds the necessary Liveblocks types to define the structure of the data used in the application, such as presence, storage, user meta, room events, thread metadata, and room info. This improves the codebase by providing better type checking and autocompletion support.

It also updates the npm dependencies in the `package.json` file. It adds the `@liveblocks/client` and `@liveblocks/react` packages, as well as the `fabric` package. It updates the `next` package to version 14.2.5 and adds the `uuid` package. Additionally, it updates the dev dependencies, including `eslint` and `eslint-config-next`, and updates `typescript` to version 5.